gpu: Move DescriptorManager to gpu_resources
And put it in gpu namespace, since it is used by both
debug printf and GPU AV
