+++ BUILD.gn
@@ -190,7 +190,8 @@ vvl_sources = [
   "layers/gpu/resources/gpuav_descriptor_set.h",
   "layers/gpu/resources/gpuav_subclasses.cpp",
   "layers/gpu/resources/gpuav_subclasses.h",
-  "layers/gpu/resources/gpuav_resources.h",
+  "layers/gpu/resources/gpu_resources.h",
+  "layers/gpu/resources/gpu_resources.cpp",
   "layers/gpu/spirv/function_basic_block.cpp",
   "layers/gpu/spirv/function_basic_block.h",
   "layers/gpu/spirv/instruction.cpp",
