+++ layers/gpu/instrumentation/gpuav_shader_instrumentor.cpp
@@ -20,113 +20,7 @@
 #include "gpu/core/gpuav_state_tracker.h"
 #include "chassis/chassis_modification_state.h"
 
-// Implementation for Descriptor Set Manager class
-DescriptorSetManager::DescriptorSetManager(VkDevice device, uint32_t num_bindings_in_set)
-    : device(device), num_bindings_in_set(num_bindings_in_set) {}
-
-DescriptorSetManager::~DescriptorSetManager() {
-    for (auto &pool : desc_pool_map_) {
-        DispatchDestroyDescriptorPool(device, pool.first, nullptr);
-    }
-    desc_pool_map_.clear();
-}
-
-VkResult DescriptorSetManager::GetDescriptorSet(VkDescriptorPool *out_desc_pool, VkDescriptorSetLayout ds_layout,
-                                                VkDescriptorSet *out_desc_sets) {
-    std::vector<VkDescriptorSet> desc_sets;
-    VkResult result = GetDescriptorSets(1, out_desc_pool, ds_layout, &desc_sets);
-    assert(result == VK_SUCCESS);
-    if (result == VK_SUCCESS) {
-        *out_desc_sets = desc_sets[0];
-    }
-    return result;
-}
-
-VkResult DescriptorSetManager::GetDescriptorSets(uint32_t count, VkDescriptorPool *out_pool, VkDescriptorSetLayout ds_layout,
-                                                 std::vector<VkDescriptorSet> *out_desc_sets) {
-    auto guard = Lock();
-    const uint32_t default_pool_size = kItemsPerChunk;
-    VkResult result = VK_SUCCESS;
-    VkDescriptorPool pool_to_use = VK_NULL_HANDLE;
-
-    assert(count > 0);
-    if (0 == count) {
-        return result;
-    }
-    out_desc_sets->clear();
-    out_desc_sets->resize(count);
-
-    for (auto &pool : desc_pool_map_) {
-        if (pool.second.used + count < pool.second.size) {
-            pool_to_use = pool.first;
-            break;
-        }
-    }
-    if (VK_NULL_HANDLE == pool_to_use) {
-        uint32_t pool_count = default_pool_size;
-        if (count > default_pool_size) {
-            pool_count = count;
-        }
-
-        // TODO: The logic to compute descriptor pool sizes should not be
-        // hardcoded like so, should be dynamic depending on the descriptor sets
-        // to be created. Not too dramatic as Vulkan will gracefully fail if there is a
-        // mismatch between this and created descriptor sets.
-        const std::array<VkDescriptorPoolSize, 2> pool_sizes = {{{
-                                                                     VK_DESCRIPTOR_TYPE_STORAGE_BUFFER,
-                                                                     pool_count * num_bindings_in_set,
-                                                                 },
-                                                                 {
-                                                                     VK_DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC,
-                                                                     pool_count * num_bindings_in_set,
-                                                                 }}};
-
-        VkDescriptorPoolCreateInfo desc_pool_info = vku::InitStructHelper();
-        desc_pool_info.flags = VK_DESCRIPTOR_POOL_CREATE_FREE_DESCRIPTOR_SET_BIT;
-        desc_pool_info.maxSets = pool_count;
-        desc_pool_info.poolSizeCount = static_cast<uint32_t>(pool_sizes.size());
-        desc_pool_info.pPoolSizes = pool_sizes.data();
-        result = DispatchCreateDescriptorPool(device, &desc_pool_info, nullptr, &pool_to_use);
-        assert(result == VK_SUCCESS);
-        if (result != VK_SUCCESS) {
-            return result;
-        }
-        desc_pool_map_[pool_to_use].size = desc_pool_info.maxSets;
-        desc_pool_map_[pool_to_use].used = 0;
-    }
-    std::vector<VkDescriptorSetLayout> desc_layouts(count, ds_layout);
-
-    VkDescriptorSetAllocateInfo alloc_info = {VK_STRUCTURE_TYPE_DESCRIPTOR_SET_ALLOCATE_INFO, nullptr, pool_to_use, count,
-                                              desc_layouts.data()};
-
-    result = DispatchAllocateDescriptorSets(device, &alloc_info, out_desc_sets->data());
-    assert(result == VK_SUCCESS);
-    if (result != VK_SUCCESS) {
-        return result;
-    }
-    *out_pool = pool_to_use;
-    desc_pool_map_[pool_to_use].used += count;
-    return result;
-}
-
-void DescriptorSetManager::PutBackDescriptorSet(VkDescriptorPool desc_pool, VkDescriptorSet desc_set) {
-    auto guard = Lock();
-    auto iter = desc_pool_map_.find(desc_pool);
-    if (iter != desc_pool_map_.end()) {
-        VkResult result = DispatchFreeDescriptorSets(device, desc_pool, 1, &desc_set);
-        assert(result == VK_SUCCESS);
-        if (result != VK_SUCCESS) {
-            return;
-        }
-        desc_pool_map_[desc_pool].used--;
-        if (0 == desc_pool_map_[desc_pool].used) {
-            DispatchDestroyDescriptorPool(device, desc_pool, nullptr);
-            desc_pool_map_.erase(desc_pool);
-        }
-    }
-    return;
-}
-
+namespace gpu {
 // Trampolines to make VMA call Dispatch for Vulkan calls
 static VKAPI_ATTR PFN_vkVoidFunction VKAPI_CALL gpuVkGetInstanceProcAddr(VkInstance inst, const char *name) {
     return DispatchGetInstanceProcAddr(inst, name);
@@ -401,7 +295,7 @@ void GpuShaderInstrumentor::CreateDevice(const VkDeviceCreateInfo *pCreateInfo, 
         return;
     }
 
-    desc_set_manager = std::make_unique<DescriptorSetManager>(device, static_cast<uint32_t>(instrumentation_bindings_.size()));
+    desc_set_manager = std::make_unique<gpu::DescriptorSetManager>(device, static_cast<uint32_t>(instrumentation_bindings_.size()));
 
     const VkDescriptorSetLayoutCreateInfo debug_desc_layout_info = {VK_STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_CREATE_INFO, nullptr, 0,
                                                                     static_cast<uint32_t>(instrumentation_bindings_.size()),
@@ -951,3 +845,5 @@ void GpuShaderInstrumentor::InternalError(LogObjectList objlist, const Location 
     // with errors making it hard to see the first error with the real source of the problem).
     ReleaseDeviceDispatchObject(this->container_type);
 }
+
+}  // namespace gpu
